Ensure non-zero return code on test failures
If we don't return results back to jest then jest will return a zero return code indicating that no errors have occured. This is problematic because it makes it impossible to know if the tests failed in CI systems and act accordingly (e.g. break a CI pipeline).

This commit ensures jest is made aware of the errors and returns a non-zero return code if a test fails. This makes it possible to utilise the jest return code in a CI system.